+++ package.json
@@ -3,8 +3,8 @@
   "version": "0.0.1",
   "description": "",
   "scripts": {
-    "local": "webpack-dev-server -r esm --config webpack.local.js & onchange '**/*.{js,jsx,html,json}' -e 'dist/**/*.*' -- prettier --write {{changed}}",
-    "build": "webpack -r esm --config"
+    "local": "webpack-dev-server --env local & onchange '**/*.{js,jsx,html,json}' -e 'dist/**/*.*' -- prettier --write {{changed}}",
+    "build": "webpack --env"
   },
   "husky": {
     "hooks": {
@@ -25,7 +25,6 @@
     "cookie-parser": "^1.4.4",
     "core-js": "^3.2.1",
     "date-fns": "^1.30.1",
-    "esm": "^3.2.25",
     "express": "^4.16.3",
     "immer": "^3.0.0",
     "isomorphic-unfetch": "^3.0.0",
